Data Structures by Nargish Gupta" to Hindi Data Structures courses (#12107)
This pull request adds a new free Data Structures course in Hindi to the free-courses-hi.md file under the Data Structures section. The course is titled Data Structures by Nargish Gupta and is available as a YouTube playlist by NG Tutorials.

This course provides in-depth coverage of data structures concepts in Hindi, making it an excellent resource for Hindi-speaking learners seeking to build strong fundamentals in data structures and algorithms.

The new entry is added in alphabetical order according to the repository guidelines.

Course link:
https://www.youtube.com/playlist?list=PLPzfPcir5uPQ6LrB421cvMvcF3EcsGrr5